style(msvc): fix remaining C4244/C4267 sites, remove suppression -- item 2 DONE
All static_cast/casting-lambda, documenting existing intentional
narrowing -- no logic changes. sm_tests unchanged at 5966/226;
[NotesLoader],[NoteData],[bms]-tagged characterization tests (647/47)
re-checked identical.

Every remaining single-site file (54 files, root src/): ActorFrame.h,
ActorMultiVertex.h, AnnouncerManager.cpp, Attack.cpp, Background.cpp,
BackgroundUtil.cpp, Character.cpp, Course.h, CourseContentsList.cpp,
CourseUtil.cpp, CryptHelpers.cpp, EnumHelper.cpp, Font.cpp,
FontCharAliases.cpp, GameConstantsAndTypes.cpp, GameState.cpp,
HighScore.cpp, IniFile.cpp, InputMapper.cpp, MemoryCardManager.cpp,
MenuTimer.cpp, ModelTypes.cpp, MsdFile.cpp, NoteData.h (§5-protected),
NotesLoaderDWI.cpp/NotesLoaderKSF.cpp/NotesLoaderSM.cpp (§5-protected),
NotesWriterSM.cpp/NotesWriterSSC.cpp (§5-protected), OptionRow.h,
PrefsManager.cpp, Profile.h (§5-protected), RageFileDriver.cpp,
RageFileDriverDirectHelpers.cpp, RageSoundMixBuffer.h,
RageSoundReader_Chain.h, RageSoundReader_Merge.cpp,
RageSoundReader_Vorbisfile.cpp, RageSoundUtil.cpp,
RageSurface_Load_XPM.cpp, RageTextureManager.cpp, RageUtil_FileDB.cpp,
ScreenDemonstration.cpp, ScreenDimensions.cpp, ScreenEnding.cpp,
ScreenEvaluation.cpp, ScreenHighScores.cpp,
ScreenOptionsManageCourses.cpp, ScreenOptionsManageEditSteps.cpp,
ScreenTextEntry.cpp, TrailUtil.cpp, Tween.cpp, WheelBase.h,
WorkoutGraph.cpp -- same recurring narrowing patterns throughout
(size_t/double/lua_Integer-family values into int/unsigned/float/
lua_Number, iterator-diffs, reverse-loop bounds, RString::Left/Right
lengths, Lua bindings).

Then, discovered right at the finish line, a fourth measurement gap:
every prior 'clean, complete' rebuild this sweep had silently stopped
at the alphabetical end of root src/*.cpp and never reached
src/arch/ or src/archutils/Win32/ at all -- a genuinely incomplete
build every time, invisible because the log tail always looked
plausible. Separately, the dedup regex never matched toolchain paths
with spaces/parens (Program Files (x86)), hiding a genuine site
attributed to <algorithm>'s std::transform internals (instantiated
from Song.cpp's ::tolower calls).

Fixed: Song.cpp's two transform(...,::tolower) sites (casting lambda,
also fixes the classic tolower-with-negative-char UB); 14 files / 25
sites across src/arch/ and src/archutils/Win32/:
RageSoundDriver_WDMKS.cpp (5), ErrorStrings.cpp (4),
MovieTexture_FFMpeg.cpp (3), CrashHandlerNetworking.cpp (2),
InputHandler_DirectInput.cpp (2), ArchHooks_Win32.cpp,
DSoundHelpers.cpp, DialogUtil.cpp, InputHandler_Win32_RTIO.cpp,
MemoryCardDriverThreaded_Windows.cpp,
RageSoundDriver_Generic_Software.cpp, RegistryAccess.cpp, USB.cpp,
VideoDriverInfo.cpp (1 each) -- mostly Win32 API int/DWORD/WORD params
fed size_t.

One site was in vendored code (extern/ffmpeg-w32-prebuilt's
libavutil/common.h, instantiated via MovieTexture_FFMpeg.cpp) -- not
patched (never edit vendored trees, same policy as the ixwebsocket
clang-tidy exclusion). Scoped an MSVC-only #pragma warning(push) /
disable: 4244 / pop around just the FFmpeg #include block in
MovieTexture_FFMpeg.h instead.

A genuinely edit-free, full-tree --clean-first Debug rebuild now
completes with exit code 0 and zero C4244/C4267 warnings anywhere.
/wd4244 /wd4267 removed from src/CMakeLists.txt permanently -- the
MSVC-warning ratchet is complete: all 5 categories (C4189, C4702,
C4100, C4244, C4267) are now promoted to -Werror. Verified with a full
clean Release rebuild (also zero warnings) + --SelfTest green.

+++ src/ActorMultiVertex.h
@@ -123,7 +123,7 @@ public:
 		RectF rect;
 		float delay;
 	};
-	int GetNumStates() const override { return _states.size(); }
+	int GetNumStates() const override { return static_cast<int>(_states.size()); }
 	void AddState(const State& new_state) { _states.push_back(new_state); }
 	void RemoveState(std::size_t i)
 	{ ASSERT(i < _states.size()); _states.erase(_states.begin()+i); }
