+++ src/MenuTimer.cpp
@@ -103,7 +103,7 @@ void MenuTimer::Update( float fDeltaTime )
 		SOUND->PlayOnceFromAnnouncer( "hurry up" );
 
 
-	int iCrossed = std::floor(fOldSecondsLeft);
+	int iCrossed = static_cast<int>(std::floor(fOldSecondsLeft));
 	if( fOldSecondsLeft > iCrossed && fNewSecondsLeft < iCrossed )	// crossed
 	{
 		if( iCrossed <= WARNING_START )
