+++ src/RageSoundReader_Merge.cpp
@@ -194,7 +194,7 @@ int RageSoundReader_Merge::Read( float *pBuffer, int iFrames )
 		/* GetNextSourceFrame for each active sound should be the same.  If any differ,
 		 * delay the later sounds until the earlier ones catch back up to put them
 		 * back in sync. */
-		int iEarliestSound = distance( aNextSourceFrames.begin(), min_element( aNextSourceFrames.begin(), aNextSourceFrames.end() ) );
+		int iEarliestSound = static_cast<int>(distance( aNextSourceFrames.begin(), min_element( aNextSourceFrames.begin(), aNextSourceFrames.end() ) ));
 
 		/* Normally, m_iNextSourceFrame should already be aligned with the GetNextSourceFrame of our
 		 * sounds.  If it's not, adjust it and return. */
